poprawki światła

+++ VertexClass.cs
@@ -49,26 +49,26 @@ namespace GK_Proj_2
             PvVecAfter = matrix.Transform(PvVec);
         }
 
-        public (int ymin, int ymax, int xmin, double m) GetEdgeAtrib(Vertex vother)
+        public (double ymin, double ymax, double xmin, double m) GetEdgeAtrib(Vertex vother)
         {
-            int xmin, ymin, ymax, xmax;
+            double xmin, ymin, ymax, xmax;
             double m;
             if (vother.pointAfter.Y > pointAfter.Y)
             {
-                ymin = (int)pointAfter.Y;
-                xmin = (int)pointAfter.X;
-                ymax = (int)vother.pointAfter.Y;
-                xmax = (int)vother.pointAfter.X;
+                ymin = pointAfter.Y;
+                xmin = pointAfter.X;
+                ymax = vother.pointAfter.Y;
+                xmax = vother.pointAfter.X;
             }
             else
             {
-                ymin = (int)vother.pointAfter.Y;
-                xmin = (int)vother.pointAfter.X;
-                ymax = (int)pointAfter.Y;
-                xmax = (int)pointAfter.X;
+                ymin = vother.pointAfter.Y;
+                xmin = vother.pointAfter.X;
+                ymax = pointAfter.Y;
+                xmax = pointAfter.X;
             }
 
-            m = ymax == ymin ? 0 : (double)(xmax - xmin) / (double)(ymax - ymin);
+            m = ymax == ymin ? 0 : (xmax - xmin) / (ymax - ymin);
 
             return (ymin, ymax, xmin, m);
         }
